@@ -33,7 +33,7 @@ const config = {
   k8sApiServerUrl_remoteenvs: `${k8sServerUrl}/apis/applicationconnector.kyma-project.io/v1alpha1/remoteenvironments/`,
   k8sDashboardApiUrl: `${k8sServerUrl}/api/v1/namespaces/kube-system/services/kubernetes-dashboard/proxy/api/v1/`,
   k8sApiServerUrl_servicecatalog: `${k8sServerUrl}/apis/servicecatalog.k8s.io/v1beta1/`,
-  k8sApiServerUrl_ui: `${k8sServerUrl}/apis/ui.kyma.cx/v1alpha1/`,
+  k8sApiServerUrl_ui: `${k8sServerUrl}/apis/ui.kyma-project.io/v1alpha1/`,
   k8sApiServerUrl_rbac: `${k8sServerUrl}/apis/rbac.authorization.k8s.io/v1/`,
   k8sDashboardUrl: `${k8sServerUrl}/namespaces/kube-system/services/kubernetes-dashboard/proxy`,
   k8sServerUrl,
