@@ -1,4 +1,4 @@
-import { ElementRef } from '@angular/core';
+import { ElementRef, EventEmitter } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { LabelsInputComponent } from './labels-input.component';
@@ -24,6 +24,21 @@ describe('LabelsInputComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+    expect(component.labelsChangeEmitter$).toEqual(new EventEmitter());
+  });
+
+  describe('NgOnInit', () => {
+    it('initializes labels to empty array if labels has falsy value', () => {
+      component.labels = null;
+      component.ngOnInit();
+      expect(component.labels).toEqual([]);
+    });
+
+    it('does not change labels value if it is truthy', () => {
+      component.labels = ['k1:v1'];
+      component.ngOnInit();
+      expect(component.labels).toEqual(['k1:v1']);
+    });
   });
 
   describe('validateNewLabel()', () => {
@@ -53,6 +68,7 @@ describe('LabelsInputComponent', () => {
     beforeEach(() => {
       component.wrongLabelMessage = '';
       component.newLabel = 'any:label';
+      spyOn(component, 'validateNewLabel');
     });
 
     it('does not update labels if no label input content', () => {
@@ -66,6 +82,11 @@ describe('LabelsInputComponent', () => {
       component.wrongLabelMessage = 'error-info';
       component.addLabel();
       expect(component.labels).toEqual([]);
+    });
+
+    it('validates label', () => {
+      component.addLabel();
+      expect(component.validateNewLabel).toHaveBeenCalled();
     });
 
     describe('valid label', () => {
